requirements.txt
grpcio>=1.74
grpcio-tools>=1.74
grpcio-reflection>=1.74
protobuf>=6.31,<7
python-dotenv==1.0.1
openai>=1.54,<2
